Add a new ConfigParameter_monitoring class to the parameters for dv/v and attenuation
